requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --cert=None --client-cert=None --index-url=None --output-file=requirements.txt --pip-args=None pyproject.toml
#
cachetools==5.5.2
    # via google-auth
certifi==2025.8.3
    # via requests
charset-normalizer==3.4.3
    # via requests
google-api-core[grpc]==2.25.1
    # via google-cloud-pubsub
google-auth==2.40.3
    # via
    #   google-api-core
    #   google-cloud-pubsub
google-cloud-pubsub==2.31.1
    # via study-pubsub (pyproject.toml)
googleapis-common-protos[grpc]==1.70.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.14.2
    # via google-cloud-pubsub
grpcio==1.75.0
    # via
    #   google-api-core
    #   google-cloud-pubsub
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.75.0
    # via
    #   google-api-core
    #   google-cloud-pubsub
idna==3.10
    # via requests
importlib-metadata==8.7.0
    # via opentelemetry-api
loguru==0.7.3
    # via study-pubsub (pyproject.toml)
opentelemetry-api==1.37.0
    # via
    #   google-cloud-pubsub
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-sdk==1.37.0
    # via google-cloud-pubsub
opentelemetry-semantic-conventions==0.58b0
    # via opentelemetry-sdk
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-pubsub
protobuf==6.32.1
    # via
    #   google-api-core
    #   google-cloud-pubsub
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
python-dotenv==1.1.1
    # via study-pubsub (pyproject.toml)
requests==2.32.5
    # via google-api-core
rsa==4.9.1
    # via google-auth
typing-extensions==4.15.0
    # via
    #   grpcio
    #   opentelemetry-api
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
urllib3==2.5.0
    # via requests
zipp==3.23.0
    # via importlib-metadata
